Extract webpack code to configure.ts
