+++ packages/docusaurus/src/commands/build.ts
@@ -15,7 +15,8 @@ import {handleBrokenLinks} from '../server/brokenLinks';
 
 import {createBuildClientConfig} from '../webpack/client';
 import createServerConfig from '../webpack/server';
-import {executePluginsConfigureWebpack, compile} from '../webpack/utils';
+import {executePluginsConfigureWebpack} from '../webpack/configure';
+import {compile} from '../webpack/utils';
 import {PerfLogger} from '../utils';
 
 import {loadI18n} from '../server/i18n';
