+++ packages/docusaurus/src/commands/start/webpack.ts
@@ -13,11 +13,11 @@ import WebpackDevServer from 'webpack-dev-server';
 import evalSourceMapMiddleware from 'react-dev-utils/evalSourceMapMiddleware';
 import {createPollingOptions} from './watcher';
 import {
-  executePluginsConfigureWebpack,
   formatStatsErrorMessage,
   getHttpsConfig,
   printStatsWarnings,
 } from '../../webpack/utils';
+import {executePluginsConfigureWebpack} from '../../webpack/configure';
 import {createStartClientConfig} from '../../webpack/client';
 import type {StartCLIOptions} from './start';
 import type {Props} from '@docusaurus/types';
